+++ .idea/misc.xml
@@ -71,6 +71,10 @@
         <entry key="../../../../layout/compose-model-1622716659069.xml" value="0.20555555555555555" />
         <entry key="../../../../layout/compose-model-1622716751444.xml" value="0.20555555555555555" />
         <entry key="../../../../layout/compose-model-1622717563443.xml" value="2.0" />
+        <entry key="../../../../layout/compose-model-1623292226754.xml" value="0.19537037037037036" />
+        <entry key="../../../../layout/compose-model-1623292342895.xml" value="2.0" />
+        <entry key="../../../../layout/compose-model-1623292387885.xml" value="0.17609797297297297" />
+        <entry key="../../../../layout/compose-model-1623294042850.xml" value="0.1935185185185185" />
         <entry key="app/src/main/res/font/font.xml" value="0.3041025641025641" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.15925925925925927" />
       </map>
