+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
 
 
     <application
-        android:name=".MyApplication"
+        android:name=".CowinAlertApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.CowinAlert">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.CowinAlert.NoActionBar">
