--- app/src/main/java/com/santoshpillai/cowinalert/Navigation.kt
@@ -1,1 +1,0 @@
-package com.santoshpillai.cowinalert
