+++ app/src/main/java/com/santoshpillai/cowinalert/data/local/AlertDatabase.kt
@@ -1,9 +1,11 @@
-package com.santoshpillai.cowinalert
+package com.santoshpillai.cowinalert.data.local
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.santoshpillai.cowinalert.data.model.Alert
+import com.santoshpillai.cowinalert.data.model.Result
 
 @Database(entities = [Alert::class, Result::class], version = 3, exportSchema = false)
 abstract class AlertDatabase : RoomDatabase() {
