+++ app/src/main/java/com/santoshpillai/cowinalert/ui/MainActivity.kt
@@ -1,20 +1,27 @@
-package com.santoshpillai.cowinalert
+package com.santoshpillai.cowinalert.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.NavType
 import androidx.navigation.compose.*
+import com.santoshpillai.cowinalert.data.local.AlertDatabase
+import com.santoshpillai.cowinalert.ui.alertscreen.AlertScreen
+import com.santoshpillai.cowinalert.ui.alertscreen.AlertViewModel
+import com.santoshpillai.cowinalert.ui.alertscreen.AlertViewModelFactory
+import com.santoshpillai.cowinalert.ui.createscreen.CreateAlertScreen
+import com.santoshpillai.cowinalert.ui.createscreen.CreateAlertViewModel
+import com.santoshpillai.cowinalert.ui.createscreen.CreateAlertViewModelFactory
+import com.santoshpillai.cowinalert.ui.resultscreen.ResultScreen
+import com.santoshpillai.cowinalert.data.model.Result
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
 
 class MainActivity : ComponentActivity() {
-    @ExperimentalAnimationApi
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val application = requireNotNull(this).application
